feat: New config, helper and routes for standard Express structure
